+++ pkg/util/reservation/reservation.go
@@ -20,7 +20,11 @@ import (
 	"fmt"
 
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/klog/v2"
+	"k8s.io/kubernetes/pkg/api/v1/resource"
+	"k8s.io/kubernetes/pkg/scheduler/framework"
 
 	"github.com/koordinator-sh/koordinator/apis/extension"
 	schedulingv1alpha1 "github.com/koordinator-sh/koordinator/apis/scheduling/v1alpha1"
@@ -164,4 +168,167 @@ func IsReservationExpired(r *schedulingv1alpha1.Reservation) bool {
 
 func GetReservationNodeName(r *schedulingv1alpha1.Reservation) string {
 	return r.Status.NodeName
+}
+
+func SetReservationExpired(r *schedulingv1alpha1.Reservation) {
+	r.Status.Phase = schedulingv1alpha1.ReservationFailed
+	// not duplicate expired info
+	idx := -1
+	isReady := false
+	for i, condition := range r.Status.Conditions {
+		if condition.Type == schedulingv1alpha1.ReservationConditionReady {
+			idx = i
+			isReady = condition.Status == schedulingv1alpha1.ConditionStatusTrue
+		}
+	}
+	if idx < 0 { // if not set condition
+		condition := schedulingv1alpha1.ReservationCondition{
+			Type:               schedulingv1alpha1.ReservationConditionReady,
+			Status:             schedulingv1alpha1.ConditionStatusFalse,
+			Reason:             schedulingv1alpha1.ReasonReservationExpired,
+			LastProbeTime:      v1.Now(),
+			LastTransitionTime: v1.Now(),
+		}
+		r.Status.Conditions = append(r.Status.Conditions, condition)
+	} else if isReady { // if was ready
+		condition := schedulingv1alpha1.ReservationCondition{
+			Type:               schedulingv1alpha1.ReservationConditionReady,
+			Status:             schedulingv1alpha1.ConditionStatusFalse,
+			Reason:             schedulingv1alpha1.ReasonReservationExpired,
+			LastProbeTime:      v1.Now(),
+			LastTransitionTime: v1.Now(),
+		}
+		r.Status.Conditions[idx] = condition
+	} else { // if already not ready
+		r.Status.Conditions[idx].Reason = schedulingv1alpha1.ReasonReservationExpired
+		r.Status.Conditions[idx].LastProbeTime = v1.Now()
+	}
+}
+
+func SetReservationSucceeded(r *schedulingv1alpha1.Reservation) {
+	r.Status.Phase = schedulingv1alpha1.ReservationSucceeded
+	idx := -1
+	for i, condition := range r.Status.Conditions {
+		if condition.Type == schedulingv1alpha1.ReservationConditionReady {
+			idx = i
+		}
+	}
+	if idx < 0 { // if not set condition
+		condition := schedulingv1alpha1.ReservationCondition{
+			Type:               schedulingv1alpha1.ReservationConditionReady,
+			Status:             schedulingv1alpha1.ConditionStatusFalse,
+			Reason:             schedulingv1alpha1.ReasonReservationSucceeded,
+			LastProbeTime:      v1.Now(),
+			LastTransitionTime: v1.Now(),
+		}
+		r.Status.Conditions = append(r.Status.Conditions, condition)
+	} else {
+		r.Status.Conditions[idx].Status = schedulingv1alpha1.ConditionStatusFalse
+		r.Status.Conditions[idx].Reason = schedulingv1alpha1.ReasonReservationSucceeded
+		r.Status.Conditions[idx].LastProbeTime = v1.Now()
+	}
+}
+
+func SetReservationAvailable(r *schedulingv1alpha1.Reservation, nodeName string) {
+	r.Status.NodeName = nodeName
+	r.Status.Phase = schedulingv1alpha1.ReservationAvailable
+	r.Status.CurrentOwners = make([]corev1.ObjectReference, 0)
+
+	requests := ReservationRequests(r)
+	r.Status.Allocatable = requests
+	r.Status.Allocated = nil
+
+	// initialize the conditions
+	r.Status.Conditions = []schedulingv1alpha1.ReservationCondition{
+		{
+			Type:               schedulingv1alpha1.ReservationConditionScheduled,
+			Status:             schedulingv1alpha1.ConditionStatusTrue,
+			Reason:             schedulingv1alpha1.ReasonReservationScheduled,
+			LastProbeTime:      v1.Now(),
+			LastTransitionTime: v1.Now(),
+		},
+		{
+			Type:               schedulingv1alpha1.ReservationConditionReady,
+			Status:             schedulingv1alpha1.ConditionStatusTrue,
+			Reason:             schedulingv1alpha1.ReasonReservationAvailable,
+			LastProbeTime:      v1.Now(),
+			LastTransitionTime: v1.Now(),
+		},
+	}
+}
+
+func ReservationRequests(r *schedulingv1alpha1.Reservation) corev1.ResourceList {
+	requests, _ := resource.PodRequestsAndLimits(&corev1.Pod{
+		Spec: r.Spec.Template.Spec,
+	})
+	return requests
+}
+
+func ReservePorts(r *schedulingv1alpha1.Reservation) framework.HostPortInfo {
+	portInfo := framework.HostPortInfo{}
+	for _, container := range r.Spec.Template.Spec.Containers {
+		for _, podPort := range container.Ports {
+			portInfo.Add(podPort.HostIP, string(podPort.Protocol), podPort.HostPort)
+		}
+	}
+	return portInfo
+}
+
+// MatchReservationOwners checks if the scheduling pod matches the reservation's owner spec.
+// `reservation.spec.owners` defines the DNF (disjunctive normal form) of ObjectReference, ControllerReference
+// (extended), LabelSelector, which means multiple selectors are firstly ANDed and secondly ORed.
+func MatchReservationOwners(pod *corev1.Pod, r *schedulingv1alpha1.Reservation) bool {
+	// assert pod != nil && r != nil
+	// Owners == nil matches nothing, while Owners = [{}] matches everything
+	for _, owner := range r.Spec.Owners {
+		if MatchObjectRef(pod, owner.Object) &&
+			MatchReservationControllerReference(pod, owner.Controller) &&
+			matchLabelSelector(pod, owner.LabelSelector) {
+			return true
+		}
+	}
+	return false
+}
+
+func MatchObjectRef(pod *corev1.Pod, objRef *corev1.ObjectReference) bool {
+	// `ResourceVersion`, `FieldPath` are ignored.
+	// since only pod type are compared, `Kind` field is also ignored.
+	return objRef == nil ||
+		(len(objRef.UID) == 0 || pod.UID == objRef.UID) &&
+			(len(objRef.Name) == 0 || pod.Name == objRef.Name) &&
+			(len(objRef.Namespace) == 0 || pod.Namespace == objRef.Namespace) &&
+			(len(objRef.APIVersion) == 0 || pod.APIVersion == objRef.APIVersion)
+}
+
+func MatchReservationControllerReference(pod *corev1.Pod, controllerRef *schedulingv1alpha1.ReservationControllerReference) bool {
+	// controllerRef matched if any of pod owner references matches the controllerRef;
+	// typically a pod has only one controllerRef
+	if controllerRef == nil {
+		return true
+	}
+	if len(controllerRef.Namespace) > 0 && controllerRef.Namespace != pod.Namespace { // namespace field is extended
+		return false
+	}
+	// currently `BlockOwnerDeletion` is ignored
+	for _, podOwner := range pod.OwnerReferences {
+		if (controllerRef.Controller == nil || podOwner.Controller != nil && *controllerRef.Controller == *podOwner.Controller) &&
+			(len(controllerRef.UID) == 0 || controllerRef.UID == podOwner.UID) &&
+			(len(controllerRef.Name) == 0 || controllerRef.Name == podOwner.Name) &&
+			(len(controllerRef.Kind) == 0 || controllerRef.Kind == podOwner.Kind) &&
+			(len(controllerRef.APIVersion) == 0 || controllerRef.APIVersion == podOwner.APIVersion) {
+			return true
+		}
+	}
+	return false
+}
+
+func matchLabelSelector(pod *corev1.Pod, labelSelector *v1.LabelSelector) bool {
+	if labelSelector == nil {
+		return true
+	}
+	selector, err := v1.LabelSelectorAsSelector(labelSelector)
+	if err != nil {
+		return false
+	}
+	return selector.Matches(labels.Set(pod.Labels))
 }
